pyproject.toml
[project]
name = "githubintegration"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=44.0.2",
    "fastapi>=0.115.12",
    "granian>=2.2.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
]
